Split station fragment in two
