+++ app/src/main/java/com/antyzero/smoksmog/ui/screen/FragmentComponent.java
@@ -1,6 +1,7 @@
 package com.antyzero.smoksmog.ui.screen;
 
 import com.antyzero.smoksmog.ui.screen.settings.GeneralSettingsFragment;
+import com.antyzero.smoksmog.ui.screen.start.fragment.NetworkStationFragment;
 
 import dagger.Subcomponent;
 
@@ -12,4 +13,6 @@ import dagger.Subcomponent;
 public interface FragmentComponent {
 
     void inject( GeneralSettingsFragment generalSettingsFragment );
+
+    void inject( NetworkStationFragment networkStationFragment );
 }
