+++ app/src/main/java/com/antyzero/smoksmog/ui/screen/start/fragment/LocationStationFragmentComponent.java
@@ -1,8 +1,9 @@
-package com.antyzero.smoksmog.ui.screen.start;
+package com.antyzero.smoksmog.ui.screen.start.fragment;
 
 
 import com.antyzero.smoksmog.google.GoogleModule;
 import com.antyzero.smoksmog.ui.screen.FragmentModule;
+import com.antyzero.smoksmog.ui.screen.start.fragment.StationFragment;
 
 import dagger.Subcomponent;
 
@@ -12,7 +13,7 @@ import dagger.Subcomponent;
                 GoogleModule.class
         }
 )
-public interface StationFragmentComponent {
+public interface LocationStationFragmentComponent {
 
-    void inject( StationFragment stationFragment );
+    void inject( LocationStationFragment fragment );
 }
